% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccuCor2_V1.0.R
\name{dual_correction}
\alias{dual_correction}
\title{Natural Abundance correction for 13C-15N or 13C-2H tracer labeling data}
\usage{
dual_correction(
  InputFile,
  InputSheetName,
  MetaboliteListName,
  Isotopes,
  Resolution = 1e+05,
  Autofill = F,
  CarbonNaturalAbundance = c(0.9893, 0.0107),
  HydrogenNaturalAbundance = c(0.999885, 0.000115),
  NitrogenNaturalAbundance = c(0.99636, 0.00364),
  OxygenNaturalAbundance = c(0.99757, 0.00038, 0.00205),
  SulfurNaturalAbundance = c(0.95, 0.0075, 0.0425),
  SiliconNaturalAbundance = c(0.92223, 0.04685, 0.03092),
  ChlorineNaturalAbundance = c(0.7576, 0.2424),
  BromineNaturalAbundance = c(0.5069, 0.4931),
  C13Purity = 0.99,
  H2N15Purity = 0.99,
  ResDefAt = 200,
  output_base = NULL,
  output_filetype = "xlsx"
)
}
\arguments{
\item{InputFile}{String representing the name of the Input file}

\item{InputSheetName}{String representing the name of excel sheet that contains the input data}

\item{MetaboliteListName}{String representing the name of the excel database the name, formula and charge information of each metabolite}

\item{Isotopes}{String that specify the type of tracer isotopes, "CN" or "CH"}

\item{Resolution}{For Exactive, the Resolution is 100000, defined at Mw 200}

\item{Autofill}{default: FALSE. If the intensity of an isotopologue is not provided, should 0 be assumed?}

\item{CarbonNaturalAbundance}{vector of C Natural Abundance in the order of (12C,13C)}

\item{HydrogenNaturalAbundance}{vector of H Natural Abundance in the order of (1H,2H)}

\item{NitrogenNaturalAbundance}{vector of N Natural Abundance in the order of (14N,15N)}

\item{OxygenNaturalAbundance}{vector of O Natural Abundance in the order of (??, ??, ??)}

\item{SulfurNaturalAbundance}{vector of S Natural Abundance in the order of (32S,33S,34S)}

\item{SiliconNaturalAbundance}{vector of Si Natural Abundance in the order of (28Si,29Si,30Si)}

\item{ChlorineNaturalAbundance}{vector of Cl Natural Abundance in the order of (35Cl,37Cl)}

\item{BromineNaturalAbundance}{vector of Br Natural Abundance in the order of (79Br,81Br)}

\item{C13Purity}{default:0.99.The isotopic purity for C13 tracer.}

\item{H2N15Purity}{default:0.99. The isotopic purity for H2/N15 tracer.}

\item{ResDefAt}{Resolution defined at (in Mw), e.g. 200 Mw}

\item{output_base}{Path to basename of output file, default is the basename
of the input path. `_corrected` will be appended. If `FALSE` then no
output file is written.}

\item{output_filetype}{Filetype of the output file, one of: 'xls', xlsx',
'csv', or 'tsv'. The default is 'xlsx'.}
}
\value{
New excel sheet named: 'Corrected', 'Normalized', 'Pool size' added to the original excel file.
}
\description{
Natural Abundance correction for 13C-15N or 13C-2H tracer labeling data
}
